last commit before menu

+++ header/random.hpp
@@ -1,4 +1,4 @@
 #pragma once
 #include <stdlib.h>
-float randi(int, int);
+int randi(int, int);
 float randf(float, float);
